requirements.txt
amqp==5.3.1
Appium-Python-Client==4.3.0
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
decorator==5.1.1
Django==5.1.4
django-channels==0.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
idna==3.10
jsonpath-ng==1.7.0
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
oauthlib==3.2.2
outcome==1.3.0.post0
pbr==6.1.0
ply==3.11
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
robotframework==7.1.1
robotframework-jsonlibrary==0.5
robotframework-requests==0.9.7
rpds-py==0.21.0
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
